pyproject.toml
[tool.poetry]
name = "restaurant-api"
version = "0.1.0"
description = ""
authors = ["Keyproco <oussama.keddar@epitech.eu>"]
readme = "README.md"
packages = [{include = "restaurant_api"}]

[tool.poetry.dependencies]
python = "^3.11"
flask-smorest = "^0.40.0"
sqlalchemy = "^2.0.6"
flask-sqlalchemy = "^3.0.3"
opentelemetry-api = "1.3.0"
opentelemetry-sdk = "1.3.0"
opentelemetry-semantic-conventions = "0.22b0"
[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"